Move TypeDeserializer into TypeModule
