+++ src/main/java/org/embulk/util/config/modules/TypeModule.java
@@ -17,11 +17,18 @@
 package org.embulk.util.config.modules;
 
 import com.fasterxml.jackson.core.JsonGenerator;
+import com.fasterxml.jackson.databind.DeserializationContext;
+import com.fasterxml.jackson.databind.JsonMappingException;
 import com.fasterxml.jackson.databind.JsonSerializer;
 import com.fasterxml.jackson.databind.SerializerProvider;
+import com.fasterxml.jackson.databind.deser.std.FromStringDeserializer;
 import com.fasterxml.jackson.databind.module.SimpleModule;
 import java.io.IOException;
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.Map;
 import org.embulk.spi.type.Type;
+import org.embulk.spi.type.Types;
 
 public final class TypeModule extends SimpleModule {
     public TypeModule() {
@@ -38,5 +45,34 @@ public final class TypeModule extends SimpleModule {
                 throws IOException {
             jsonGenerator.writeString(value.getName());
         }
+    }
+
+    private static class TypeDeserializer extends FromStringDeserializer<Type> {
+        public TypeDeserializer() {
+            super(Type.class);
+        }
+
+        @Override
+        protected Type _deserialize(final String value, final DeserializationContext context) throws IOException {
+            final Type type = STRING_TO_TYPE.get(value);
+            if (type == null) {
+                throw new JsonMappingException(String.format("Unknown type name '%s'. Supported types are: %s",
+                        value, String.join(", ", STRING_TO_TYPE.keySet())));
+            }
+            return type;
+        }
+
+        static {
+            final HashMap<String, Type> builder = new HashMap<>();
+            builder.put(Types.BOOLEAN.getName(), Types.BOOLEAN);
+            builder.put(Types.LONG.getName(), Types.LONG);
+            builder.put(Types.DOUBLE.getName(), Types.DOUBLE);
+            builder.put(Types.STRING.getName(), Types.STRING);
+            builder.put(Types.TIMESTAMP.getName(), Types.TIMESTAMP);
+            builder.put(Types.JSON.getName(), Types.JSON);
+            STRING_TO_TYPE = Collections.unmodifiableMap(builder);
+        }
+
+        private static final Map<String, Type> STRING_TO_TYPE;
     }
 }
